Use "size_t" for page_size instead of "long"

+++ SoC/MT6735/dump/dump.c
@@ -44,7 +44,7 @@ int main(int argc, char *argv[])
 		return -1;
 	}
 
-	long page_size = sysconf(_SC_PAGESIZE);
+	size_t page_size = sysconf(_SC_PAGESIZE);
 	reg_offset = address % page_size;
 	reg_base = address - reg_offset;
 	pages = ((reg_offset + length + (page_size - 1)) / page_size);
